Add a new index.js allowing custom servers to call a getCounts method
This allows hooking into an existing server by just adding a simple
route instead of running an entire new server.
