+++ package.json
@@ -2,7 +2,7 @@
   "name": "shariff-backend-node",
   "version": "1.0.7",
   "description": "",
-  "main": "server.js",
+  "main": "index.js",
   "scripts": {
     "start": "node run.js",
     "test": "echo \"Error: no test specified\" && exit 1",
