+++ server.js
@@ -1,6 +1,7 @@
 var Hapi    = require('hapi');
 var Promise = require('promise');
 var config  = require('./shariff.json');
+var Shariff = require('./index.js');
 
 var server = new Hapi.Server(config.port, config.host, {
     cache: {
@@ -8,78 +9,12 @@ var server = new Hapi.Server(config.port, config.host, {
     }
 });
 
-function toJSON(data) {
-    return new Promise(function(resolve, reject) {
-
-        try {
-            var jsonString = JSON.stringify(data);
-            if (jsonString && typeof jsonString === "string") {
-                resolve(jsonString);
-            } else {
-                throw "Failed generating JSON string.";
-            }
-        }
-        catch(e) {
-            reject(Error(e));
-        }
-    });
-}
-
 server.method('getJSONOutput', function(resourceURL, next) {
-
-    var services = [
-        require('./lib/twitter'),
-        require('./lib/facebook'),
-        require('./lib/googleplus')
-    ];
-
-    var requests = services.map(function(service) {
-        return new Promise(function(resolve, reject) {
-
-            service.request(resourceURL, function(error, response, body) {
-                if ( !error && response.statusCode === 200 ) {
-
-                    if ( typeof body === "object" ) {
-                        resolve(body);
-                    } else {
-                        resolve(JSON.parse(body));
-                    }
-
-                } else {
-                    reject(error);
-                }
-            });
-        });
+    Shariff.getCounts(resourceURL).then(function(counts) {
+        next(null, counts);
+    }, function(err) {
+        next(err);
     });
-
-    // res: Array of result objects
-    function extractCounts(res) {
-        return new Promise(function(resolve, reject) {
-            var result = {};
-
-            var i = 0;
-            res.forEach(function() {
-                var count = services[i].extractCount(res[i]);
-                result[ services[i].name ] = count;
-                i++;
-            });
-
-            resolve(result);
-        });
-    }
-
-    Promise
-        .all(requests)
-        .catch(function(error) {
-            console.log(require('util').inspect(error));
-            next(error);
-        })
-        .then(extractCounts)
-        .then(toJSON)
-        .then(function(jsonOutput) {
-            next(null, jsonOutput);
-        });
-
 }, {
     cache: {
         expiresIn: config.cache.expiresIn
